Move _cast_params to commons.
