+++ st2common/st2common/models/utils/action_param_utils.py
@@ -13,7 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import ast
 import copy
+import json
+import six
+
+from st2common import log as logging
+from st2common.util import action_db as action_db_util
+
+
+LOG = logging.getLogger(__name__)
 
 
 def _merge_param_meta_values(action_meta=None, runner_meta=None):
@@ -70,3 +79,51 @@ def get_params_view(action_db=None, runner_db=None, merged_only=False):
     immutable_params = {k: merged_params[k] for k in immutable}
 
     return (required_params, optional_params, immutable_params)
+
+
+def cast_params(action_ref, params):
+    """
+    """
+    def cast_object(x):
+        if isinstance(x, str) or isinstance(x, unicode):
+            try:
+                return json.loads(x)
+            except:
+                return ast.literal_eval(x)
+        else:
+            return x
+
+    casts = {
+        'array': (lambda x: ast.literal_eval(x) if isinstance(x, str) or isinstance(x, unicode)
+                  else x),
+        'boolean': (lambda x: ast.literal_eval(x.capitalize())
+                    if isinstance(x, str) or isinstance(x, unicode) else x),
+        'integer': int,
+        'number': float,
+        'object': cast_object,
+        'string': str
+    }
+    action_db = action_db_util.get_action_by_ref(action_ref)
+    action_parameters_schema = action_db.parameters
+    runnertype_db = action_db_util.get_runnertype_by_name(action_db.runner_type['name'])
+    runner_parameters_schema = runnertype_db.runner_parameters
+    # combine into 1 list of parameter schemas
+    parameters_schema = {}
+    if runner_parameters_schema:
+        parameters_schema.update(runner_parameters_schema)
+    if action_parameters_schema:
+        parameters_schema.update(action_parameters_schema)
+    # cast each param individually
+    for k, v in six.iteritems(params):
+        parameter_schema = parameters_schema.get(k, None)
+        if not parameter_schema:
+            continue
+        parameter_type = parameter_schema.get('type', None)
+        if not parameter_type:
+            continue
+        cast = casts.get(parameter_type, None)
+        LOG.debug('Casting param: %s of type %s to type: %s', v, type(v), parameter_type)
+        if not cast:
+            continue
+        params[k] = cast(v)
+    return params
